% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{exported_from}
\alias{exported_from}
\title{Extract export lines from NAMESPACE}
\usage{
exported_from(package)
}
\arguments{
\item{package}{A string}
}
\value{
A string vector of \code{export()} lines in NAMESPACE
}
\description{
Extract export lines from NAMESPACE
}
\examples{
exported_from('readr')
}

