% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{extract_exports}
\alias{extract_exports}
\title{Extract export lines from NAMESPACE}
\usage{
extract_exports(package)
}
\arguments{
\item{package}{A string}
}
\value{
\code{export()} lines in NAMESPACE
}
\description{
Extract export lines from NAMESPACE
}
\examples{
extract_exports('readr')
}

