% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{import_export_all}
\alias{import_export_all}
\title{Write import() and export() to NAMESPACE}
\usage{
import_export_all(packages, path = ".")
}
\arguments{
\item{packages}{A string vector}

\item{path}{A string}
}
\description{
Write import() and export() to NAMESPACE
}

