% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{load_in_namespace}
\alias{load_in_namespace}
\title{Load packages and put all objects into a namespace}
\usage{
load_in_namespace(namespace, packages)
}
\arguments{
\item{namespace}{A string}

\item{packages}{A string vector}
}
\description{
Load packages and put all objects into a namespace
}
\examples{
load_in_namespace('had', c('readr', 'stringr', 'tidyr', 'dplyr', 'ggplot2'))
}

