% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{namaespace}
\alias{namaespace}
\alias{namaespace-package}
\title{namaespace: A package to create dummy packages for namespace}
\description{
namaespace: A package to create dummy packages for namespace
}

