% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{pkgs}
\alias{pkgs}
\alias{pkgs_bio}
\alias{pkgs_had}
\title{Examples of package set}
\format{An object of class \code{character} of length 6.}
\usage{
pkgs_had

pkgs_bio
}
\description{
Examples of package set

Hadley Wickham packages

BioConductor basic packages
}
\examples{
\dontrun{
load_in_namespace('had', pkgs_had)
}
\dontrun{
load_in_namespace('bio', pkgs_bio)
}
}
\keyword{datasets}

