% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs.R
\docType{data}
\name{pkgs}
\alias{pkgs}
\alias{pkgs_bio}
\alias{pkgs_tidyverse}
\title{Examples of package set}
\format{An object of class \code{character} of length 6.}
\usage{
pkgs_tidyverse

pkgs_bio
}
\description{
Examples of package set

tidyverse packages

BioConductor basic packages
}
\examples{
\dontrun{
load_in_namespace('tdy', pkgs_tidyverse)
}
\dontrun{
load_in_namespace('bio', pkgs_bio)
}
}
\keyword{datasets}

