% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{pkgs}
\alias{pkgs}
\alias{pkgs_bio}
\alias{pkgs_had}
\title{Examples of package set}
\format{\preformatted{ chr [1:6] "readr" "stringr" "tidyr" "plyr" "dplyr" "ggplot2"
}}
\usage{
pkgs_had

pkgs_bio
}
\description{
Examples of package set

Hadley Wickham packages

BioConductor basic packages
}
\examples{
load_in_namespace('had', pkgs_had)
#load_in_namespace('bio', pkgs_bio)
}
\keyword{datasets}

